[tests] refactor: consolidate MiniMax H3 dummy data
Combine the T2VA, FL2VA, and Ref2VA parquet generators in one task-aware helper while preserving their schemas and deterministic image bytes. Update the runner and GPU-smoke selector to use the consolidated module.

AI assistance (pi coding agent) was used for this change.
